promjena tipa promjenljive za cijenu iz int u double

+++ src/ba/unsa/etf/rpr/Artikl.java
@@ -2,7 +2,7 @@ package ba.unsa.etf.rpr;
 
 public class Artikl {
     private String naziv;
-    private int cijena;
+    private double cijena;
     private String kod;
 
     public Artikl(String naziv, int cijena, String kod) {
